requirements.txt
# Core LangGraph and LangChain dependencies
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0

# MCP (Model Context Protocol) adapters
langchain-mcp-adapters>=0.1.0

# FastAPI and web server
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Vector database and embeddings
pinecone>=5.0.0
openai>=1.0.0

# Environment and configuration
python-dotenv>=1.0.0

# HTTP requests and async support
httpx>=0.25.0
aiohttp>=3.9.0

# JSON and data handling
pydantic>=2.5.0

# Logging and utilities
structlog>=23.2.0

# Development and testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
